+++ includes/PF_FormPrinter.php
@@ -849,8 +849,6 @@ END;
 						if ( $cur_value !== '' &&
 							( $form_field->hasFieldArg( 'mapping template' ) ||
 							$form_field->hasFieldArg( 'mapping property' ) ||
-							( $form_field->hasFieldArg( 'mapping cargo table' ) &&
-							$form_field->hasFieldArg( 'mapping cargo field' ) ) ||
 							$form_field->getUseDisplayTitle() ) ) {
 							// If the input type is "tokens', the value is not
 							// an array, but the delimiter still needs to be set.