+++ includes/PF_Template.php
@@ -1,8 +1,7 @@
 <?php
 /**
  * Defines a class, PFTemplate, that represents a MediaWiki "infobox"
- * template that holds structured data, which may or may not be
- * additionally stored by Cargo and/or Semantic MediaWiki.
+ * template that holds structured data, which may be stored by Semantic MediaWiki.
  *
  * @author Yaron Koren
  * @ingroup PF
@@ -17,7 +16,6 @@ class PFTemplate {
 	private $mTemplateParams;
 	private $mConnectingProperty;
 	private $mCategoryName;
-	private $mCargoTable;
 	private $mAggregatingProperty;
 	private $mAggregationLabel;
 	private $mTemplateFormat;
@@ -264,13 +262,6 @@ class PFTemplate {
 		$this->mCategoryName = $categoryName;
 	}
 
-	// @codeCoverageIgnoreStart
-	public function setCargoTable( $cargoTable ) {
-		$this->mCargoTable = str_replace( ' ', '_', $cargoTable );
-	}
-
-	// @codeCoverageIgnoreEnd
-
 	public function setFullWikiTextStatus( $status ) {
 		$this->mFullWikiText = $status;
 	}
@@ -312,18 +303,12 @@ END;
 				}
 				$text .= "|" . $field->getFieldName() . "=\n";
 			}
-			// @codeCoverageIgnoreStart
-			$cargoInUse = false;
-			$cargoDeclareCall = '';
-			$cargoStoreCall = '';
-			// @codeCoverageIgnoreEnd
-
 			$templateFooter = wfMessage( 'pf_template_docufooter' )->inContentLanguage()->text();
 			$text .= <<<END
 }}
 </pre>
 $templateFooter
-$cargoDeclareCall</noinclude><includeonly>$cargoStoreCall
+</noinclude><includeonly>
 END;
 		} else {
 			$text = <<<END
@@ -339,15 +324,9 @@ END;
 				}
 				$text .= $field->toWikitext();
 			}
-			// @codeCoverageIgnoreStart
-			$cargoInUse = false;
-			$cargoDeclareCall = '';
-			$cargoStoreCall = '';
-			// @codeCoverageIgnoreEnd
-
 			$text .= <<<END
 }}
-$cargoDeclareCall</noinclude><includeonly>$cargoStoreCall
+</noinclude><includeonly>
 END;
 			// @codeCoverageIgnoreStart
 			// This branch is only reachable when SMW is not installed.
@@ -471,14 +450,7 @@ END;
 				}
 			}
 
-			// @codeCoverageIgnoreStart
-			// If we're using Cargo, fields can simply be displayed
-			// normally - no need for any special tags - *unless*
-			// the field holds a list of Page values, in which case
-			// we need to apply #arraymap.
-			$isCargoListOfPages = $cargoInUse && $field->isList() && $field->getFieldType() == 'Page';
-			// @codeCoverageIgnoreEnd
-			if ( !$fieldProperty && !$isCargoListOfPages ) {
+			if ( !$fieldProperty ) {
 				if ( $separator != '' ) {
 					$tableText .= "$separator ";
 				}
@@ -581,10 +553,7 @@ END;
 			$text .= "$fieldStart ";
 		}
 
-		// @codeCoverageIgnoreStart
-		$cargoInUse = false;
-		// @codeCoverageIgnoreEnd
-		$text .= $field->createText( $cargoInUse );
+		$text .= $field->createText();
 
 		$fieldEnd = $this->mFieldEnd;
 		MediaWikiServices::getInstance()->getHookContainer()->run(